Comment: 1. Remove isShowEmptyView -> tablewView backgroundView = EmptyView
